Modificado onHitStart da linha temporal para usar Timer ao invés de teste se posição da linha chegou ao fim: Isso melhorou a performance pois nao é necessário testar o se chegou ao fim a cada quadro. Também melhorou a sensação da música, os intervalos são mais precisos

+++ app/src/main/java/com/viana/soundprogramming/sound/Sound.kt
@@ -11,7 +11,7 @@ class Sound(private val soundId: Int) {
     fun play() {
         timer.schedule(object : TimerTask() {
             override fun run() {
-                SoundManager.instance.play(soundId, volume.toFloat())
+                SoundManager.instance.play(soundId, volume)
                 ProgrammingVibrator.vibrate((volume * 10).toLong())
             }
         }, delayMillis)
